+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2025-05-20
+
+### Added
+- Install generator (`rails g solidstats:install`) to automatically mount routes
+- Rake task (`rake solidstats:install`) as an alias for the generator
+- Improved documentation for installation process
+
 ## [0.0.2] - 2025-05-19
 
 ### Added